1404: Use correct value for default value behavior
The supported values are FIELD_BEHAVIOR_DEFAULT and FIELD_BEHAVIOR_NONE
https://api.drupal.org/api/drupal/modules%21field%21field.api.php/function/hook_field_widget_info/7.x

Took my a while to figure out what this setting actually does. The
default behavior is to embed the widget in the instance setting for the
field, so that administraors can set the default values there. Since we
provide our own interface for setting default values and use these
settings in the widget, the appropiate vaule must be NONE. e.g. we don't
support the default value setting from Field UI.